requirements.txt
PySide6==6.7.2
PyMuPDF==1.24.10
Pillow==10.4.0
pytest==8.3.2
pyinstaller==6.10.0
ruff==0.6.9
